add some utils for server-action functions feat: integrated change-password logic. Refactor server actions

+++ src/actions/auth/log-out-account.ts
@@ -1,39 +1,32 @@
 "use server";
 import { userServiceConfig } from "#/configs/user-service.app-config";
-import { notLoggedErrorTxt } from "#/constants/frequent-errors-from-client";
-import { internalErrTxt } from "#/integration/constants/messages-from-services";
 import { sessionAuthenticator_S_A } from "#/integration/user-service/auth/cookie-authenticator.integrator";
 import { mainUserServiceFetcher } from "#/integration/user-service/user-service-fetcher.integrator-util";
 import type { ServerActionResponseWithPromise } from "#T/integrator-main-types";
 import type { Profile_ResponseTypes } from "#user-service/shared/response-patterns/profile.routes.js";
 import { cookies } from "next/headers";
+import { userServiceRawResponsePreHandler } from "../server-actions-utils/user-service-raw-response-pre-handler";
+import { authenticator_ServerActionResponsePreHandler } from "../server-actions-utils/authenticator-server-action-response-pre-handler";
 
 export async function LogOutAccount_ServerAction(): ServerActionResponseWithPromise {
     const _cookies = await cookies();
-    const auth = await sessionAuthenticator_S_A();
-    if (!auth) {
-        return { errors: [notLoggedErrorTxt], ok: false };
+    const userSession = authenticator_ServerActionResponsePreHandler(
+        await sessionAuthenticator_S_A(),
+    );
+    if (userSession.errors || !userSession.auth) {
+        return userSession.errors;
     }
-    if (auth === 500) {
-        return { errors: [internalErrTxt], ok: false };
-    }
-
     const res =
         await mainUserServiceFetcher<Profile_ResponseTypes.delete_avatar>({
             url: `/v1/authentication/logout`,
             method: "DELETE",
-            ip: auth.ip,
-            agent: auth.agent,
-            session_token: auth.data.session.token,
+            ip: userSession.auth.ip,
+            agent: userSession.auth.agent,
+            session_token: userSession.auth.data.session.token,
         });
-
-    if (!res || res === 500) {
-        return { errors: [internalErrTxt], ok: false };
-    }
-
-    if (res.data) {
-        _cookies.delete(userServiceConfig.session_token_name);
-        return { msg: res.message, ok: true };
-    }
-    return { errors: res.errors || ["default"], ok: false };
+    return await userServiceRawResponsePreHandler(res, {
+        onSuccessFunction: () => {
+            _cookies.delete(userServiceConfig.session_token_name);
+        },
+    });
 }
